Добавил работу с клиентом. Изменил структуру БД - промежуточный коммит

+++ app/routes.py
@@ -20,6 +20,7 @@ from app.models import Role
 from app.models import Task
 from app.models import Mode
 from app.models import File
+from app.models import Status
 from flask_login import current_user
 from flask_login import login_user
 from flask_login import logout_user
@@ -57,7 +58,6 @@ def login():
         if not next_page or url_parse(next_page).netloc != '':
             next_page = url_for('index')
         return redirect(next_page)
-
     return render_template('login.html',  title="Вход", form=form)
 
 
@@ -90,7 +90,7 @@ def register():
 @app.route('/add_task', methods=['GET', 'POST'])
 @login_required
 def add_task():
-    path_upload = os.path.join(app.config['UPLOADED_PATH'], "new_task")
+    path_upload = os.path.join(app.config['UPLOADED_PATH'], "tmp", current_user.username, "new_task")
 
     # загрузка файлов из dropzone
     if request.method == 'POST':
@@ -108,6 +108,8 @@ def add_task():
             task = Task(comment=form.comment.data)
             task.mode = Mode.query.get(int(mode))
             task.author = current_user
+
+            task.task_status = Status.query.filter_by(name='queue').first()
             db.session.add(task)
             db.session.commit()
 
@@ -115,17 +117,22 @@ def add_task():
             # для записи в папку используем iso 8601 фрмат
             dir_name = datetime.now().isoformat()
             path_to_files = os.path.join(os.path.realpath(app.config['UPLOADED_PATH']), "tasks", dir_name)
-            shutil.copytree(path_upload, path_to_files)
 
-            for file_name in os.listdir(path_to_files):
-                file = File()
-                file.add(path_to_files, file_name)
-                db.session.add(file)
-                db.session.commit()
-                task.files_id.append(file)
-                db.session.commit()
+            if os.path.isdir(path_upload):
+                shutil.copytree(path_upload, path_to_files)
 
-        shutil.rmtree(path_upload)
+            if os.path.isdir(path_to_files):
+                for file_name in os.listdir(path_to_files):
+                    file = File()
+                    file.add(path_to_files, file_name)
+                    db.session.add(file)
+                    db.session.commit()
+                    task.files_id.append(file)
+                    db.session.commit()
+
+        if os.path.isdir(path_upload):
+            shutil.rmtree(path_upload)
+
         flash("Успешная регистрация задания")
         return redirect(url_for('index'))
